Changed size types from int to off_t for large files + implemented first UI for progress tracking

+++ includes/Browser.hpp
@@ -62,7 +62,7 @@ public:
     Gtk::TreeModelColumn<Glib::ustring> name;
     Gtk::TreeModelColumn<Glib::ustring> type;
     Gtk::TreeModelColumn<int> hardlinks;
-    Gtk::TreeModelColumn<int> size;
+    Gtk::TreeModelColumn<off_t> size;
     Gtk::TreeModelColumn<Glib::ustring> access;
     Gtk::TreeModelColumn<std::time_t> lastEdited;
   };
